% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCanonical3SS.R
\name{getCanonical3SS}
\alias{getCanonical3SS}
\title{Get locations of the first five AG 3'Splice site motifs}
\usage{
getCanonical3SS(ag)
}
\arguments{
\item{ag}{Vector of distances to the AG splice site motif.}
}
\value{
Locations of the first five AG dinucleotides
}
\description{
Takes a variable length vector of distances to the AG motif,
sorts and returns the first five.
If there are less than five elements in the vector, returns the sorted vector
and fills the remainder of the values with 300.
}
\examples{
sequence <- paste(c("A","T","C","G")[sample(1:4, 501, replace=TRUE)] , collapse="")
AG_seq_positions <- gregexpr("AG",sequence, 252,501)
canon_hits <- getCanonical3SS(AG_seq_positions[[1]])
}
\author{
Beth Signal
}

