% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQueryLoc.R
\name{getExonDists}
\alias{getExonDists}
\title{Get the closest 3' and 5' exons}
\usage{
getExonDists(queryLine, exons, queryType)
}
\arguments{
\item{queryLine}{line from a query GenomicRanges}

\item{exons}{GenomicRanges containing exon co-ordinates.
Should be produced by gtfToExons()}

\item{queryType}{type of query. "SNP" or "region"}
}
\value{
GenomicRanges with distance to the closest 3' and 5' exons,
whether these exons are part of the same gene (i.e is the location intronic),
and the identifiers for the 3' and 5' exons.
}
\description{
Finds the closest annotated exons from a genomic co-ordinate.
Returns the distance to the 3' exon, distance to the 5' exon,
ids of the 3' and 5' exon,
and if the exons are from the same parent gene
}
\author{
Beth Signal
}
\keyword{internal}
