% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictBranchpoints.R
\name{getPPT}
\alias{getPPT}
\title{Get the best polypyramidine tract}
\usage{
getPPT(attributes)
}
\arguments{
\item{attributes}{query attributes GenomicRanges}
}
\value{
distance to the start of the longest PPT, and its length
}
\description{
Takes a query genomic sequence, finds all potential polypyramidine tracts (PPTs)
between the test site and the annotated 3'exon.
Returns the distance to the start of the longest PPT, and its length.
}
\author{
Beth Signal
}
\keyword{internal}
