% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfToExons.R
\name{gtfToExons}
\alias{gtfToExons}
\title{Convert GTF file to exon location file}
\usage{
gtfToExons(gtf)
}
\arguments{
\item{gtf}{file containing the gtf annotation.}
}
\value{
exon annotation GRanges
}
\description{
Converts a GTF annotation to exon locations
}
\examples{
smallExons <- system.file("extdata","gencode.v24.annotation.small.gtf",
package = "branchpointer")
exons <- gtfToExons(smallExons)
}
\author{
Beth Signal
}
