% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRegions.R
\name{makeRegions}
\alias{makeRegions}
\title{Make branchpoint window regions}
\usage{
makeRegions(id, idType, exons)
}
\arguments{
\item{id}{identifier for the query gene/transcript/exon id}

\item{idType}{type of id to match in the exon annotation file (\code{"gene_id"},
\code{"transcript_id"}, or \code{"exon_id"})}

\item{exons}{data.frame containing exon co-ordinates.
Should be produced by gtfToExons()}
}
\value{
Granges with formatted query
}
\description{
Genrate branchpoint window regions corresponding to annotated exon(s) within a
queried gene, transcript or exon id
}
\examples{
smallExons <- system.file("extdata","gencode.v24.annotation.small.gtf",package = "branchpointer")
exons <- gtfToExons(smallExons)
windowquery <- makeRegions("ENSG00000139618", "gene_id", exons)
windowquery <- makeRegions("ENST00000357654", "transcript_id", exons)
windowquery <- makeRegions("ENSE00003518965", "exon_id", exons)
}
\author{
Beth Signal
}
