% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBranchpointWindow.R
\name{plotStructure}
\alias{plotStructure}
\title{Plots transcript structures}
\usage{
plotStructure(exonID, exons, keepTranscripts = "overlapping")
}
\arguments{
\item{exonID}{id of the exon to plot}

\item{exons}{Granges containing exon co-ordinates.}

\item{keepTranscripts}{which transcripts to plot (\code{"overlapping"} or \code{"withExon"})
\code{"overlapping"} will plot all transcripts overlapping the exon, 
whereas \code{"withExon"} will plot all transcripts containing the exon.}
}
\value{
ggplot2 plot transcript structures
}
\description{
Plots transcript structures
}
\examples{
smallExons <- system.file("extdata","gencode.v24.annotation.small.gtf",
package = "branchpointer")
exons <- gtfToExons(smallExons)
plotStructure(exonID = "ENSE00001184784.4", exons)
}
\author{
Beth Signal
}
