% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictionsToStats.R
\name{predictionsToStats}
\alias{predictionsToStats}
\title{Convert SNP branchpoint predictions across the branchpoint window to an intronic summary}
\usage{
predictionsToStats(query, predictions, probabilityCutoff = 0.5,
  probabilityChange = 0.2)
}
\arguments{
\item{query}{query GRanges containing all SNP ids to be summarised}

\item{predictions}{site-wide branchpoint proability predictions
produced from predictBranchpoints()}

\item{probabilityCutoff}{Value to be used as the cutoff for
discriminating branchpoint sites from non-branchpoint sites
(default = \code{0.5})}

\item{probabilityChange}{Minimum probability score change
required to call a branchpoint site as deleted or created by
a SNP (default = \code{0.2})}
}
\value{
GRanges with summarised branchpoint changes
occuring within the intron due to a SNP.
}
\description{
Takes predictions of branchpoint probabilities from a reference
and alternative SNP and summarises the effect(s) of the SNP.
}
\examples{
smallExons <- system.file("extdata","gencode.v24.annotation.small.gtf",
package = "branchpointer")
exons <- gtfToExons(smallExons)
genome <- BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38

querySNP <- system.file("extdata","SNP_example.txt", package = "branchpointer")
query <- readQueryFile(querySNP,queryType = "SNP")
query <- getQueryLoc(query,queryType = "SNP",exons = exons, filter = FALSE)
predictions <- predictBranchpoints(query,queryType = "SNP",BSgenome = genome)
snpStats <- predictionsToStats(query,predictions)
}
\author{
Beth Signal
}
