% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpToQuery.R
\name{snpToQuery}
\alias{snpToQuery}
\title{Gets a branchpointer formatted query from refsnp ids}
\usage{
snpToQuery(refSNP, mart.snp)
}
\arguments{
\item{refSNP}{Vector of refsnp ids}

\item{mart.snp}{biomaRt mart object specifying the BioMart database and dataset to be used}
}
\value{
formatted SNP query data.frame
}
\description{
Searches Biomart for refsnp ids, and pulls genomic location and sequence identity information
Reformats alleles so each query has only one alternative allele
}
\examples{
mart.snp <- biomaRt::useMart("ENSEMBL_MART_SNP", dataset="hsapiens_snp", host="www.ensembl.org")
query <- snpToQuery("rs17000647", mart.snp)
}
\author{
Beth Signal
}
