% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQueryLoc.R
\name{exonsToIntrons}
\alias{exonsToIntrons}
\title{Convert exon annotation GRanges to intron locations}
\usage{
exonsToIntrons(exons, maxDist = 50)
}
\arguments{
\item{exons}{data.frame containing exon co-ordinates.
Should be produced by gtfToExons()}

\item{maxDist}{Maximum distance from the 3' exon to create the branchpoint region.}
}
\value{
vector of chromosome names and intronic locations
}
\description{
Converts exon annotation to intron locations overlapping the branchpoint region
for exculsion of non-branchpoint region SNPs
Returns a character vector of chromosome locations
}
\author{
Beth Signal
}
\keyword{internal}
