% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBranchpointSequence.R
\name{getBranchpointSequence}
\alias{getBranchpointSequence}
\title{Get branchpoint sequence features
Gets intronic sequence covering the branchpoint window and extracts predictive features}
\usage{
getBranchpointSequence(query, unique_id = "test", query_type,
  working_directory = ".", genome = NA, rm_chr = FALSE, bedtools_location,
  use_parallel = FALSE, cores = 1, useBSgenome = FALSE, BSgenome = NULL)
}
\arguments{
\item{query}{branchpointer query data.frame
must have chromosome at position 2, genomic co-ordinate at position 3,
and strand at position 4.}

\item{unique_id}{unique string identifier for intermediate .bed and .fa files.}

\item{query_type}{type of branchpointer query. "SNP" or "region".}

\item{working_directory}{directory where intermediate .bed and .fa are located}

\item{genome}{.fa genome file location}

\item{rm_chr}{remove "chr" before chromosome names before writing bed file.
Required if genome sequence names do not contain "chr"}

\item{bedtools_location}{bedtools binary location (which bedtools)}

\item{use_parallel}{use parallelisation to speed up code?}

\item{cores}{number of cores to use in parallelisation (default = \code{1})}

\item{useBSgenome}{Use a BSgenome object for sequence retrieval?  (default = \code{1})
Overridden if a genome .fa and a bedtools_location are sepcified.}

\item{BSgenome}{BSgenome object}
}
\value{
data.frame with all features required to predict branchpoint probability scores
}
\description{
Get branchpoint sequence features
Gets intronic sequence covering the branchpoint window and extracts predictive features
}
\examples{
small_exons <- system.file("extdata","gencode.v24.annotation.exons.small.txt",
package = "branchpointer")
exons <- readExonAnnotation(small_exons)
genome <- BSgenome.Hsapiens.UCSC.hg38::BSgenome.Hsapiens.UCSC.hg38

query_snp <- system.file("extdata","SNP_example.txt", package = "branchpointer")
query <- readQueryFile(query_snp,query_type = "SNP")
query <- getQueryLoc(query,query_type = "SNP",exons = exons, filter = FALSE)
query_attributes <- getBranchpointSequence(query,
query_type = "SNP",
useBSgenome = TRUE,
BSgenome = genome)

query <- makeRegions("ENSE00003541068.1", "exon_id", exons)
query_attributes <- getBranchpointSequence(query,
query_type = "region",
useBSgenome = TRUE,
BSgenome = genome)
}
\author{
Beth Signal
}

