% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQueryLoc.R
\name{getQueryLoc}
\alias{getQueryLoc}
\title{Find the closest 3' and 5' exons to a branchpointer query}
\usage{
getQueryLoc(query, queryType, maxDist = 50, filter = TRUE, exons,
  useParallel = FALSE, cores = 1)
}
\arguments{
\item{query}{branchpointer query GenomicRanges
must have chromosome at position 2, genomic co-ordinate at position 3,
and strand at position 4.}

\item{queryType}{type of query file (\code{"SNP"} or \code{"region"})}

\item{maxDist}{maximum distance a SNP can be from an annotated 3' exon.}

\item{filter}{remove SNP queries prior to finding finding nearest exons.}

\item{exons}{data.frame containing exon co-ordinates.
Should be produced by gtfToExons()}

\item{useParallel}{use parallelisation to speed up code?
(reccomended for > 500 query entries)}

\item{cores}{number of cores to use in parallelisation (default = \code{1})}
}
\value{
GenomicRanges with the query and its location relative to the 3' and 5' exons
}
\description{
Finds the closest annotated exons from genomic co-ordinates in a branchpointer query GRanges
}
\examples{
smallExons <- system.file("extdata","gencode.v24.annotation.small.gtf",package = "branchpointer")
exons <- gtfToExons(smallExons)

querySNP <- system.file("extdata","SNP_example.txt", package = "branchpointer")
query <- readQueryFile(querySNP,queryType = "SNP")
query <- getQueryLoc(query,queryType = "SNP",exons = exons, filter = FALSE)
}
\author{
Beth Signal
}
