% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readQueryFile.R
\name{readQueryFile}
\alias{readQueryFile}
\title{Read a query file}
\usage{
readQueryFile(queryFile, queryType)
}
\arguments{
\item{queryFile}{tab delimited file containing query information.
For intronic regions should be in the format: region id, chromosome name, region start, region id, strand.
For SNP variants should be in the format: SNP id, chromosome name, SNP position, strand, reference allele (A/T/C/G), alternative allele (A/T/C/G)}

\item{queryType}{type of query file (\code{"SNP"} or \code{"region"})}
}
\value{
data.frame with formatted query
}
\description{
Reads and formats a query file, generated from snpToQuery, or manually generated.
Converts unstranded SNPs to two entries for each strand.
Checks for duplicate names and replaces if found.
}
\examples{
querySNP <- system.file("extdata","SNP_example.txt", package = "branchpointer")
query <- readQueryFile(querySNP,queryType = "SNP")

queryIntron <- system.file("extdata","intron_example.txt", package = "branchpointer")
query <- readQueryFile(queryIntron,queryType = "region")
}
\author{
Beth Signal
}
